pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "python-multipart>=0.0.20",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
